requirements.txt
gradio==4.44.0
openai==1.12.0
anthropic==0.8.1
google-generativeai==0.3.2
groq==0.4.1
python-dotenv==1.0.1
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.17.0
PyPDF2==3.0.1
python-docx==1.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
requests==2.31.0
pushover-complete==1.2.0
schedule==1.2.0
typing-extensions==4.8.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
psutil>=5.8.0
